Core framework files updated
* Event and Logger uses configuration to act/deact
* Process: now has sanity check [need to add more]
* Support 404 pages
* Support fully open (no login required) website
* Request and Session cleans-up, adjustments
* Response: Add ability to handle PetakUmpet 'system' Views

+++ Logger.php
@@ -3,22 +3,29 @@
 namespace PetakUmpet;
 
 class Logger {
+  const NONE  = 0;
   const INFO  = 1;
   const DEBUG = 2;
 
   public static function log($message='', $level=self::INFO)
   {
     $marker = array(
-      self::INFO => 'PetakUmpet INFO: ',
-      self::DEBUG => 'PetakUmpet DEBUG: ',
+      self::INFO => '[PetakUmpet INFO ] ',
+      self::DEBUG => '[PetakUmpet DEBUG] ',
     );
 
     $request = Singleton::acquire('\\PetakUmpet\\Request');
+    $config = Singleton::acquire('\\PetakUmpet\\Config');
 
     $logfile = PU_DIR . DS . 'app' . DS . $request->getApplication() . DS . 'res' . DS . 'log' . DS . 'app.log';
     $curtime = new \DateTime;
 
-    file_put_contents($logfile, '['. $curtime->format('Y-m-d H:i:s') . ']' . $marker[$level] . $message . "\n", FILE_APPEND | LOCK_EX);
+    if ($config->getLogLevel() >= $level) {
+      if (!is_file($logfile) || !is_writable($logfile)) {
+        die ("Logger Error, have you setup logfile and access mode correctly?");
+      }
+      file_put_contents($logfile, '['. $curtime->format('Y-m-d H:i:s') . ']' . $marker[$level] . $message . "\n", FILE_APPEND | LOCK_EX);
+    }
   }
   
 }
